requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
pyotp==2.9.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
starlette==0.27.0
itsdangerous==2.1.2 